Restrict access to admin only pages
Every page set that could allow access to the database or allow getting
it is now only available to admins. This includes users and groups which
are admin stuff. Probes and graphs allow to store/execute SQL queries
that could be used to leak/destroy information if accessed by
unprivileged users.